enhancement: move query logic into the component
